PID in double, MCS removed comments, timer fixed peripheral inits for encoder

+++ src/motion/PID.h
@@ -11,12 +11,12 @@
 #include "stm32g431xx.h"
 
 class PID {
-  uint32_t kp, ki, kd;
+  double kp, ki, kd;
   int32_t min, max;
 
   int64_t integral_max, integral_min;
 
-  int64_t comp_proportional, comp_integral, comp_derivative;
+  double comp_proportional, comp_integral, comp_derivative;
 
   int32_t error, last_error, derivative_error, last_setpoint;
 
@@ -24,15 +24,15 @@ public:
   PID();
   void reset();
 
-  void set_coefficients(float new_kp, float new_ki, float new_kd, uint32_t new_freq);
+  void set_coefficients(double new_kp, double new_ki, double new_kd, uint32_t new_freq);
 
-  void set_kp(float new_kp);
-  void set_ki(float new_ki, uint32_t new_freq);
-  void set_kd(float new_kd, uint32_t new_freq);
+  void set_kp(double new_kp);
+  void set_ki(double new_ki, uint32_t new_freq);
+  void set_kd(double new_kd, uint32_t new_freq);
   void set_output_limit(int32_t new_limits);
   void set_anti_windup(int32_t new_limit);
 
-  void get_coefficients(float* ret_kp, float* ret_ki, float* ret_kd);
+  void get_coefficients(double* ret_kp, double* ret_ki, double* ret_kd);
 
   int16_t compute(int32_t input, int32_t setpoint);
